Upgrade non-compatibility sahders in Basic_viewer to glsl 150 and add a context creation before the creation of the QApp in draw_xxx examples
